backend/requirements.txt
weaviate-client==4.17.0
sentence-transformers==5.1.1
tqdm==4.67.1
biopython==1.85
python-dotenv==1.1.1
fastapi==0.104.1
uvicorn==0.24.0
cors==0.0.1